requirements.txt
flask
mcstatus
GitPython